refactor env vars + add federationConfig.json
