+++ src/lib/federation.ts
@@ -1,0 +1,4 @@
+import { createConfig } from '@lawallet/utils';
+import fedConfig from '~/constants/federationConfig.json';
+
+export const federationConfig = createConfig(fedConfig);
